refactor: move project lists to separate components
