+++ src/components/project-lists/node-project-list/node-project-list.module.scss
@@ -1,0 +1,11 @@
+@use '../../../styles/main.scss' as *;
+
+.projects_nodejs {
+  @include flex($direction: column);
+  gap: $gap-s;
+  max-width: 1600px;
+  overflow-y: auto;
+  max-height: calc(100vh - 220px);
+  overflow-x: hidden;
+
+}
